// src/gen-ts/src/models/Warehouse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Company Management API
 * API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux, mouvements financiers, dépenses, achats et frais de transport. Basée sur le schéma de base de données fourni.
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime'
import type { CrupdateJob } from './CrupdateJob'
import {
  CrupdateJobFromJSON,
  CrupdateJobFromJSONTyped,
  CrupdateJobToJSON,
  CrupdateJobToJSONTyped,
} from './CrupdateJob'
import type { AuditUser } from './AuditUser'
import {
  AuditUserFromJSON,
  AuditUserFromJSONTyped,
  AuditUserToJSON,
  AuditUserToJSONTyped,
} from './AuditUser'

/**
 *
 * @export
 * @interface Warehouse
 */
export interface Warehouse {
  /**
   *
   * @type {string}
   * @memberof Warehouse
   */
  comment?: string
  /**
   *
   * @type {Date}
   * @memberof Warehouse
   */
  created_at?: Date
  /**
   *
   * @type {Date}
   * @memberof Warehouse
   */
  updated_at?: Date
  /**
   *
   * @type {AuditUser}
   * @memberof Warehouse
   */
  created_by?: AuditUser
  /**
   *
   * @type {AuditUser}
   * @memberof Warehouse
   */
  updated_by?: AuditUser
  /**
   *
   * @type {string}
   * @memberof Warehouse
   */
  id?: string
  /**
   *
   * @type {string}
   * @memberof Warehouse
   */
  name?: string
  /**
   *
   * @type {string}
   * @memberof Warehouse
   */
  description?: string
  /**
   *
   * @type {CrupdateJob}
   * @memberof Warehouse
   */
  job?: CrupdateJob
}

/**
 * Check if a given object implements the Warehouse interface.
 */
export function instanceOfWarehouse(value: object): value is Warehouse {
  return true
}

export function WarehouseFromJSON(json: any): Warehouse {
  return WarehouseFromJSONTyped(json, false)
}

export function WarehouseFromJSONTyped(json: any, ignoreDiscriminator: boolean): Warehouse {
  if (json == null) {
    return json
  }
  return {
    comment: json['comment'] == null ? undefined : json['comment'],
    created_at: json['created_at'] == null ? undefined : new Date(json['created_at']),
    updated_at: json['updated_at'] == null ? undefined : new Date(json['updated_at']),
    created_by: json['created_by'] == null ? undefined : AuditUserFromJSON(json['created_by']),
    updated_by: json['updated_by'] == null ? undefined : AuditUserFromJSON(json['updated_by']),
    id: json['id'] == null ? undefined : json['id'],
    name: json['name'] == null ? undefined : json['name'],
    description: json['description'] == null ? undefined : json['description'],
    job: json['job'] == null ? undefined : CrupdateJobFromJSON(json['job']),
  }
}

export function WarehouseToJSON(json: any): Warehouse {
  return WarehouseToJSONTyped(json, false)
}

export function WarehouseToJSONTyped(
  value?: Warehouse | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value
  }

  return {
    comment: value['comment'],
    created_at:
      value['created_at'] == null ? value['created_at'] : value['created_at'].toISOString(),
    updated_at:
      value['updated_at'] == null ? value['updated_at'] : value['updated_at'].toISOString(),
    created_by: AuditUserToJSON(value['created_by']),
    updated_by: AuditUserToJSON(value['updated_by']),
    id: value['id'],
    name: value['name'],
    description: value['description'],
    job: CrupdateJobToJSON(value['job']),
  }
}